requirements.txt
fastapi
uvicorn
requests
PyPDF2
langchain
google-auth
google-auth-oauthlib
google-api-python-client
sentence-transformers
faiss-cpu
ibm-watson
pyyaml
pydantic
python-dotenv
